+++ x-pack/plugins/synthetics/public/apps/synthetics/components/monitor_details/monitor_status/monitor_status_data.ts
@@ -79,7 +79,7 @@ export function getColorBands(euiTheme: EuiThemeComputed, colorMode: EuiThemeCol
 }
 
 export function getSkippedVizColor(euiTheme: EuiThemeComputed) {
-  return euiTheme.colors.body;
+  return euiTheme.colors.lightestShade;
 }
 
 export function getErrorVizColor(euiTheme: EuiThemeComputed) {
